Improve LoginVC
The UITextFields are replaced with the new LoginTextField class. The property to loginButton is removed since a reference to it is not needed. The shake feature is utilized when login/signup fails. The view now contains a header label. The appearance of the text fields are updated to look nicer.